requirements.txt
httpx>=0.27
pydantic>=2.5
PyYAML>=6.0
mss>=9.0
Pillow>=10.0
tzlocal>=5.2
pystray>=0.19
opencv-python>=4.8

